+++ src/renderer/components/basemap/WMXLayerTable.js
